+++ src/WindowsMcp/Program.cs
@@ -71,6 +71,7 @@ internal static class Program
         builder.Services.AddSingleton<IProcessService, ProcessService>();
         builder.Services.AddSingleton<IWindowService, WindowService>();
         builder.Services.AddSingleton<IWmiService, WmiService>();
+        builder.Services.AddSingleton<IStorageService, StorageService>();
         builder.Services.AddSingleton<IEnvService, EnvService>();
         builder.Services.AddSingleton<IPowerService, PowerService>();
         builder.Services.AddSingleton<INotificationService, NotificationService>();